refactor: 패키지 위치 변경 -> naver 패키지 추가
